refactor: 지하철 이용요금 기능구현
- 요금 클래스 분리
- 계산 로직 요금 클래스로 이관
